requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.6
numpy==1.26.4
parso==0.8.3
pexpect==4.9.0
prettyprint==0.1.5
prompt-toolkit==3.0.43
ptyprocess==0.7.0
PuLP==2.8.0
pure-eval==0.2.2
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
PyYAML==6.0.1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
traitlets==5.14.2
typing_extensions==4.10.0
wcwidth==0.2.13
